requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.55.0
Pillow==11.0.0
pydantic==2.10.2
